+++ src/pages/404.jsx
@@ -1,16 +1,15 @@
 import { Link } from 'gatsby';
-import PropTypes from 'prop-types';
+// import PropTypes from 'prop-types';
 import React from 'react';
 
 import SEO from '../views/components/SEO';
 import MainLayout from '../views/layouts/MainLayout';
 
-const NotFoundPage = ({ location }) => (
+const NotFoundPage = () => (
     <MainLayout>
         <SEO
             title='demo.p404.headerTitle'
             description='demo.p404.description'
-            location={location}
             pageType='404'
         />
         <h1>404</h1>
@@ -18,9 +17,5 @@ const NotFoundPage = ({ location }) => (
         <Link to='/'>Page 2</Link>
     </MainLayout>
 );
-
-NotFoundPage.propTypes = {
-    location: PropTypes.object.isRequired,
-};
 
 export default NotFoundPage;