+++ src/pages/page-2.jsx
@@ -1,15 +1,13 @@
 import { Link } from 'gatsby';
-import PropTypes from 'prop-types';
 import React from 'react';
 
 import SEO from '../views/components/SEO';
 import MainLayout from '../views/layouts/MainLayout';
 
-const SecondPage = ({ location }) => (
+const SecondPage = () => (
     <MainLayout>
         <SEO
             title='demo.page2.headerTitle'
-            location={location}
             description='demo.page2.description'
         />
         <div className='container'>
@@ -18,9 +16,5 @@ const SecondPage = ({ location }) => (
         </div>
     </MainLayout>
 );
-
-SecondPage.propTypes = {
-    location: PropTypes.object.isRequired,
-};
 
 export default SecondPage;