+++ src/views/layouts/Footer/Footer.jsx
@@ -1,35 +1,47 @@
 import React from 'react';
+import { useStaticQuery, graphql } from 'gatsby';
 
-import { name } from '../../../../config/siteConfig';
 import { copyright, business, author } from '../../../../config/socialConfig';
 
 import './Footer.styl';
 
-const Footer = () => (
-    <footer>
-        <div className='container'>
-            <div>
-                {copyright}
-                {` `}
-                <a
-                    href={business.networks.github}
-                    target='_blank'
-                    rel='noopener noreferrer'
-                >
-                    {name}
-                </a>
-                <br />
-                <span>by </span>
-                <a
-                    href={author.networks.github}
-                    target='_blank'
-                    rel='noopener noreferrer'
-                >
-                    {author.username}
-                </a>
+const Footer = () => {
+    const data = useStaticQuery(graphql`
+        query SiteFooterQuery {
+            site {
+                siteMetadata {
+                    name
+                }
+            }
+        }
+    `);
+
+    return (
+        <footer>
+            <div className='container'>
+                <div>
+                    {copyright}
+                    {` `}
+                    <a
+                        href={business.networks.github}
+                        target='_blank'
+                        rel='noopener noreferrer'
+                    >
+                        {data.site.siteMetadata.name}
+                    </a>
+                    <br />
+                    <span>by </span>
+                    <a
+                        href={author.networks.github}
+                        target='_blank'
+                        rel='noopener noreferrer'
+                    >
+                        {author.username}
+                    </a>
+                </div>
             </div>
-        </div>
-    </footer>
-);
+        </footer>
+    );
+};
 
 export default Footer;